group files into new pkgs
