+++ contracts/accounts.go
@@ -10,8 +10,8 @@ import (
 	"os"
 
 	"github.com/NethermindEth/juno/core/felt"
-	starknetgo "github.com/NethermindEth/starknet.go"
 	"github.com/NethermindEth/starknet.go/artifacts"
+	"github.com/NethermindEth/starknet.go/curve"
 	"github.com/NethermindEth/starknet.go/rpc"
 	"github.com/NethermindEth/starknet.go/utils"
 )
@@ -74,7 +74,7 @@ func InstallAndWaitForAccount[V *rpc.Provider](ctx context.Context, provider V, 
 		return nil, errors.New("empty account")
 	}
 	privateKeyString := fmt.Sprintf("0x%x", privateKey)
-	publicKey, _, err := starknetgo.Curve.PrivateToPoint(privateKey)
+	publicKey, _, err := curve.Curve.PrivateToPoint(privateKey)
 	if err != nil {
 		return nil, err
 	}
