--- examples/curve/go.mod
@@ -1,12 +1,0 @@
-module github.com/NethermindEth/starknet.go/examples/curve
-
-go 1.18
-
-replace github.com/NethermindEth/starknet.go => ../../
-
-require github.com/NethermindEth/starknet.go v0.3.1
-
-require (
-	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e // indirect
-	golang.org/x/sys v0.0.0-20220615213510-4f61da869c0c // indirect
-)
