--- utils_test.go
@@ -1,23 +1,0 @@
-package starknetgo
-
-import (
-	"testing"
-
-	"github.com/NethermindEth/starknet.go/utils"
-)
-
-func TestGeneral_SplitFactStr(t *testing.T) {
-	data := []map[string]string{
-		{"input": "0x3", "h": "0x0", "l": "0x3"},
-		{"input": "0x300000000000000000000000000000000", "h": "0x3", "l": "0x0"},
-	}
-	for _, d := range data {
-		l, h := utils.SplitFactStr(d["input"]) // 0x3
-		if l != d["l"] {
-			t.Errorf("expected %s, got %s", d["l"], l)
-		}
-		if h != d["h"] {
-			t.Errorf("expected %s, got %s", d["h"], h)
-		}
-	}
-}
